Condense grants into a single file
