--- packages/idos-sdk-js/src/lib/grants/index.ts
@@ -1,2 +1,0 @@
-export * from "./grant";
-export * from "./grants";
