+++ packages/idos-sdk-js/src/lib/index.ts
@@ -1,5 +1,5 @@
 export * from "./idos";
-export { default as idOSGrant } from "./grants/grant";
+export { default as idOSGrant } from "./grants";
 export * from "@idos-network/controllers/enclave";
 export * from "./utils";
 export * from "./grants";
